Fold the new crate to existing diskann-benchmark crate
