--- diskann-benchmark-multi-vector/Cargo.toml
@@ -1,30 +1,0 @@
-[package]
-name = "diskann-benchmark-multi-vector"
-version.workspace = true
-description.workspace = true
-authors.workspace = true
-documentation.workspace = true
-license.workspace = true
-edition.workspace = true
-
-[[bin]]
-name = "benchmark-multi-vector"
-path = "src/bin.rs"
-
-[dependencies]
-anyhow.workspace = true
-diskann-utils = { workspace = true, default-features = false }
-half = { workspace = true, features = ["rand_distr"] }
-diskann-benchmark-runner = { workspace = true }
-diskann-quantization = { workspace = true }
-diskann-vector = { workspace = true }
-rand.workspace = true
-serde = { workspace = true, features = ["derive"] }
-serde_json.workspace = true
-thiserror.workspace = true
-
-[lints]
-workspace = true
-
-[dev-dependencies]
-tempfile.workspace = true
