+++ diskann-benchmark/src/backend/mod.rs
@@ -7,10 +7,12 @@ mod disk_index;
 mod exhaustive;
 mod filters;
 mod index;
+mod multi_vector;
 
 pub(crate) fn register_benchmarks(registry: &mut diskann_benchmark_runner::registry::Benchmarks) {
     exhaustive::register_benchmarks(registry);
     disk_index::register_benchmarks(registry);
     index::register_benchmarks(registry);
     filters::register_benchmarks(registry);
+    multi_vector::register_benchmarks(registry);
 }
