+++ diskann-quantization/src/multi_vector/matrix.rs
@@ -244,6 +244,18 @@ pub unsafe trait NewOwned<T>: ReprOwned {
 #[derive(Debug, Clone, Copy)]
 pub struct Defaulted;
 
+/// An initializer argument to [`NewOwned`] that invokes the wrapped closure for each
+/// element.
+///
+/// # Example
+/// ```
+/// use diskann_quantization::multi_vector::{Init, Mat, Standard};
+/// let mut n = 0;
+/// let mat = Mat::new(Standard::<i32>::new(1, 4).unwrap(), Init(|| { n += 1; n })).unwrap();
+/// assert_eq!(mat.as_slice(), &[1, 2, 3, 4]);
+/// ```
+pub struct Init<F>(pub F);
+
 /// Create a new [`Mat`] cloned from a view.
 pub trait NewCloned: ReprOwned {
     /// Clone the contents behind `v`, returning a new owning [`Mat`].
@@ -511,6 +523,22 @@ where
     type Error = crate::error::Infallible;
     fn new_owned(self, _: Defaulted) -> Result<Mat<Self>, Self::Error> {
         self.new_owned(T::default())
+    }
+}
+
+// SAFETY: The implementation uses guarantees from `Box` to ensure that the pointer
+// initialized by it is non-null and properly aligned to the underlying type.
+unsafe impl<T, F> NewOwned<Init<F>> for Standard<T>
+where
+    T: Copy,
+    F: FnMut() -> T,
+{
+    type Error = crate::error::Infallible;
+    fn new_owned(self, mut init: Init<F>) -> Result<Mat<Self>, Self::Error> {
+        let b: Box<[T]> = (0..self.num_elements()).map(|_| (init.0)()).collect();
+
+        // SAFETY: By construction, `b` has length `self.num_elements()`.
+        Ok(unsafe { self.box_to_mat(b) })
     }
 }
 
@@ -1765,6 +1793,22 @@ mod tests {
                 }
             }
         }
+    }
+
+    #[test]
+    fn test_standard_new_owned_with_init() {
+        let mut counter: i32 = 0;
+        let m = Mat::new(
+            Standard::<i32>::new(2, 3).unwrap(),
+            Init(|| {
+                let v = counter;
+                counter += 1;
+                v
+            }),
+        )
+        .unwrap();
+
+        assert_eq!(m.as_slice(), &[0, 1, 2, 3, 4, 5]);
     }
 
     #[test]
